Upgrade tinymce to version 4.9.3

+++ src/core/main/ts/delete/InlineFormatDelete.ts
@@ -1,11 +1,8 @@
 /**
- * InlineFormatDelete.js
- *
- * Released under LGPL License.
- * Copyright (c) 1999-2017 Ephox Corp. All rights reserved
- *
- * License: http://www.tinymce.com/license
- * Contributing: http://www.tinymce.com/contributing
+ * Copyright (c) Tiny Technologies, Inc. All rights reserved.
+ * Licensed under the LGPL or a commercial license.
+ * For LGPL see License.txt in the project root for license information.
+ * For commercial licenses see https://www.tiny.cloud/
  */
 
 import { Arr, Fun } from '@ephox/katamari';
@@ -15,7 +12,7 @@ import DeleteElement from './DeleteElement';
 import DeleteUtils from './DeleteUtils';
 import * as ElementType from '../dom/ElementType';
 import Parents from '../dom/Parents';
-import CaretFormat from '../fmt/CaretFormat';
+import * as CaretFormat from '../fmt/CaretFormat';
 
 const getParentInlines = function (rootElm, startElm) {
   const parents = Parents.parentsAndSelf(startElm, rootElm);
@@ -52,7 +49,7 @@ const deleteCaret = function (editor, forward) {
 
   return Arr.last(parentInlines).map(function (target) {
     const fromPos = CaretPosition.fromRangeStart(editor.selection.getRng());
-    if (DeleteUtils.willDeleteLastPositionInElement(forward, fromPos, target.dom())) {
+    if (DeleteUtils.willDeleteLastPositionInElement(forward, fromPos, target.dom()) && !CaretFormat.isEmptyCaretFormatElement(target)) {
       deleteLastPosition(forward, editor, target, parentInlines);
       return true;
     } else {
